Release 1.4.0

+++ entrypoint.sh
@@ -29,6 +29,7 @@ echo "ℹ︎ SLUG is $SLUG"
 # Does it even make sense for VERSION to be editable in a workflow definition?
 if [[ -z "$VERSION" ]]; then
 	VERSION=${GITHUB_REF#refs/tags/}
+	VERSION=$(echo $VERSION | sed -e "s/^v//")
 fi
 echo "ℹ︎ VERSION is $VERSION"
 
@@ -92,7 +93,11 @@ else
 fi
 
 # Copy dotorg assets to /assets
-rsync -rc "$GITHUB_WORKSPACE/$ASSETS_DIR/" assets/ --delete
+if [[ -d "$GITHUB_WORKSPACE/$ASSETS_DIR/" ]]; then
+	rsync -rc "$GITHUB_WORKSPACE/$ASSETS_DIR/" assets/ --delete
+else
+	echo "ℹ︎ No assets directory found; skipping asset copy"
+fi
 
 # Add everything and commit to SVN
 # The force flag ensures we recurse into subdirectories even if they are already added
